+++ core/src/builder/sighash.rs
@@ -44,10 +44,10 @@ pub fn create_nofn_sighash_stream(
     _evm_address: EVMAddress,
     _recovery_taproot_address: Address<NetworkUnchecked>,
     nofn_xonly_pk: XOnlyPublicKey,
-    _user_takes_after: u32,
+    _user_takes_after: u16,
     collateral_funding_amount: Amount,
     timeout_block_count: i64,
-    max_withdrawal_time_block_count: i64,
+    max_withdrawal_time_block_count: u16,
     bridge_amount_sats: Amount,
     network: bitcoin::Network,
 ) -> impl Stream<Item = Result<TapSighash, BridgeError>> {
